✅ 🧹 Use explicit uint256 instead of uint in tests

+++ test/Uint256/BaseUint256.t.sol
@@ -15,7 +15,7 @@ contract BaseUint256Test is Test {
     }
 
     function testBaseUint256AddFuzz(uint256 x, uint256 y) public {
-        uint z;
+        uint256 z;
         assembly {
             z := add(x, y)
         }
@@ -55,7 +55,7 @@ contract BaseUint256Test is Test {
     }
 
     function testBaseUint256AndFuzz(uint256 x, uint256 y) public {
-        uint z = x & y;
+        uint256 z = x & y;
         BaseUint256 result = BaseUint256.wrap(x).and(y);
         assertEq(result.unwrap(), z);
     }
@@ -68,7 +68,7 @@ contract BaseUint256Test is Test {
     }
 
     function testBaseUint256And2Fuzz(uint256 x, uint256 y) public {
-        uint z = x & y;
+        uint256 z = x & y;
         BaseUint256 result = BaseUint256.wrap(x) & BaseUint256.wrap(y);
         assertEq(result.unwrap(), z);
     }
@@ -160,7 +160,7 @@ contract BaseUint256Test is Test {
             vm.expectRevert();
             BaseUint256.wrap(x) % BaseUint256.wrap(y);
         } else {
-            uint z = x % y;
+            uint256 z = x % y;
             BaseUint256 result = BaseUint256.wrap(x) % BaseUint256.wrap(y);
             assertEq(result.unwrap(), z);
         }
@@ -182,7 +182,7 @@ contract BaseUint256Test is Test {
     }
 
     function testBaseUint256Not() public {
-        uint hundred = 100;
+        uint256 hundred = 100;
         BaseUint256 x = BaseUint256.wrap(hundred);
         BaseUint256 result = ~x;
         uint256 expected = ~hundred;
@@ -225,7 +225,7 @@ contract BaseUint256Test is Test {
     }
 
     function testBaseUint256IncrementFuzz(uint256 x) public {
-        uint z;
+        uint256 z;
         assembly {
             z := add(x, 1)
         }
@@ -246,7 +246,7 @@ contract BaseUint256Test is Test {
     }
 
     function testBaseUint256MulFuzz(uint256 x, uint256 y) public {
-        uint z;
+        uint256 z;
         assembly {
             z := mul(x, y)
         }
@@ -267,7 +267,7 @@ contract BaseUint256Test is Test {
     }
 
     function testBaseUint256MulDecimalFuzz(uint256 x, uint256 y) public {
-        uint z;
+        uint256 z;
         assembly {
             z := div(
                 mul(x, y),
@@ -298,7 +298,7 @@ contract BaseUint256Test is Test {
         uint256 x,
         uint256 y
     ) public {
-        uint z;
+        uint256 z;
         assembly {
             z := div(
                 mul(x, y),
@@ -328,7 +328,7 @@ contract BaseUint256Test is Test {
     }
 
     function testBaseUint256DivFuzz(uint256 x, uint256 y) public {
-        uint z;
+        uint256 z;
         assembly {
             z := div(x, y)
         }
