+++ test/Uint256/USDPerQuantoUint256.t.sol
@@ -15,7 +15,7 @@ contract USDPerQuantoUint256Test is Test {
     }
 
     function testUSDPerQuantoUint256AddFuzz(uint256 x, uint256 y) public {
-        uint z;
+        uint256 z;
         assembly {
             z := add(x, y)
         }
@@ -56,7 +56,7 @@ contract USDPerQuantoUint256Test is Test {
     }
 
     function testUSDPerQuantoUint256AndFuzz(uint256 x, uint256 y) public {
-        uint z = x & y;
+        uint256 z = x & y;
         USDPerQuantoUint256 result = USDPerQuantoUint256.wrap(x).and(y);
         assertEq(result.unwrap(), z);
     }
@@ -69,7 +69,7 @@ contract USDPerQuantoUint256Test is Test {
     }
 
     function testUSDPerQuantoUint256And2Fuzz(uint256 x, uint256 y) public {
-        uint z = x & y;
+        uint256 z = x & y;
         USDPerQuantoUint256 result = USDPerQuantoUint256.wrap(x) &
             USDPerQuantoUint256.wrap(y);
         assertEq(result.unwrap(), z);
@@ -165,7 +165,7 @@ contract USDPerQuantoUint256Test is Test {
             vm.expectRevert();
             USDPerQuantoUint256.wrap(x) % USDPerQuantoUint256.wrap(y);
         } else {
-            uint z = x % y;
+            uint256 z = x % y;
             USDPerQuantoUint256 result = USDPerQuantoUint256.wrap(x) %
                 USDPerQuantoUint256.wrap(y);
             assertEq(result.unwrap(), z);
@@ -189,7 +189,7 @@ contract USDPerQuantoUint256Test is Test {
     }
 
     function testUSDPerQuantoUint256Not() public {
-        uint hundred = 100;
+        uint256 hundred = 100;
         USDPerQuantoUint256 x = USDPerQuantoUint256.wrap(hundred);
         USDPerQuantoUint256 result = ~x;
         uint256 expected = ~hundred;
@@ -234,7 +234,7 @@ contract USDPerQuantoUint256Test is Test {
     }
 
     function testUSDPerQuantoUint256IncrementFuzz(uint256 x) public {
-        uint z;
+        uint256 z;
         assembly {
             z := add(x, 1)
         }
@@ -257,7 +257,7 @@ contract USDPerQuantoUint256Test is Test {
     }
 
     function testUSDPerQuantoUint256MulFuzz(uint256 x, uint256 y) public {
-        uint z;
+        uint256 z;
         assembly {
             z := mul(x, y)
         }
@@ -281,7 +281,7 @@ contract USDPerQuantoUint256Test is Test {
         uint256 x,
         uint256 y
     ) public {
-        uint z;
+        uint256 z;
         assembly {
             z := div(
                 mul(x, y),
@@ -314,7 +314,7 @@ contract USDPerQuantoUint256Test is Test {
         uint256 x,
         uint256 y
     ) public {
-        uint z;
+        uint256 z;
         assembly {
             z := div(
                 mul(x, y),
@@ -347,7 +347,7 @@ contract USDPerQuantoUint256Test is Test {
         uint256 x,
         uint256 y
     ) public {
-        uint z;
+        uint256 z;
         assembly {
             z := div(
                 mul(x, y),
@@ -379,7 +379,7 @@ contract USDPerQuantoUint256Test is Test {
     }
 
     function testUSDPerQuantoUint256DivFuzz(uint256 x, uint256 y) public {
-        uint z;
+        uint256 z;
         assembly {
             z := div(x, y)
         }
